backend/serial-fastapi/pyproject.toml
[tool.poetry]
name = "serial-fastapi"
version = "0.1.0"
description = ""
authors = ["arx55 <164836470+arx55@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
pyserial = "^3.5"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
